feat(stageTemplate): add create stage template
