--- web/src/components/stageTemplate/CreateStageTemplate/index.js
@@ -1,2 +1,0 @@
-import StageTemplateForm from './FormWrap';
-export default StageTemplateForm;
