+++ web/src/components/stageTemplate/add/FormContent.js
