+++ web/src/components/stageTemplate/add/SelectSourceType.js
