+++ web/src/components/stageTemplate/add/index.js
@@ -1,0 +1,2 @@
+import StageTemplateForm from './Form';
+export default StageTemplateForm;
