+++ web/src/components/stageTemplate/add/validate.js
